RequestResponseMiddleware moved into shared project. HttpInstrumentation is added for StockAPI and OrderAPI.
