+++ StockAPI/StockAPI.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Common.Shared\Common.Shared.csproj" />
+    <ProjectReference Include="..\OpenTelemetry.Shared\OpenTelemetry.Shared.csproj" />
   </ItemGroup>
 
 </Project>
