+++ StockAPI/appsettings.Development.json
@@ -4,5 +4,10 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "OpenTelemetry": {
+    "ServiceName": "StockAPI",
+    "Version": "1.0.0",
+    "ActivitySourceName": "ActivitySource.StockAPI"
   }
 }
